Show authors the content path of internal links.
When an internal link is added to a list, it's description
appears like "Item Title (internal: )". Some content authors have given
feedback that displaying the item's path in this text would be useful.

Implementing an ehancement which retrieves the content path of an internal
link via its GUID, and displays that value in the area which would show
the URL if the link were external.